+++ src/Argonaut.Persistence/Models/PointOfInterestEntity.cs
@@ -8,7 +8,7 @@ namespace Argonaut.Persistence.Models
     public class PointOfInterestEntity
     {
         [Key]
-        public int PointOfInterestId { get; set; }
+        public long PointOfInterestId { get; set; }
 
         [Required]
         public string Name { get; set; }
